backend/ml/speech-to-text/whisper_backend/requirements.txt
flask
flask_cors
gunicorn
openai-whisper